+++ src/egon/data/processing/zensus_vg250/zensus_population_inside_germany.py
@@ -9,8 +9,10 @@ from sqlalchemy import (
     SmallInteger,
     String,
     func,
+    select,
 )
 from sqlalchemy.ext.declarative import declarative_base
+import geopandas as gpd
 
 from egon.data import db
 import egon.data.config
@@ -129,6 +131,92 @@ class Vg250GemPopulation(Base):
     geom = Column(Geometry(srid=3035))
 
 
+class MapZensusVg250(Base):
+    __tablename__ = "egon_map_zensus_vg250"
+    __table_args__ = {"schema": "boundaries"}
+
+    zensus_population_id = Column(Integer, primary_key=True, index=True)
+    zensus_geom = Column(Geometry("POINT", 3035))
+    vg250_municipality_id = Column(Integer)
+    vg250_nuts3 = Column(String)
+
+
+def map_zensus_vg250():
+    """Perform mapping between municipalities and zensus grid"""
+
+    MapZensusVg250.__table__.drop(bind=db.engine(), checkfirst=True)
+    MapZensusVg250.__table__.create(bind=db.engine(), checkfirst=True)
+
+    # Get information from data configuration file
+    cfg = egon.data.config.datasets()["map_zensus_vg250"]
+
+    local_engine = db.engine()
+
+    db.execute_sql(
+        f"""DELETE FROM
+        {cfg['targets']['map']['schema']}.{cfg['targets']['map']['table']}"""
+    )
+
+    gdf = db.select_geodataframe(
+        f"""SELECT * FROM
+        {cfg['sources']['zensus_population']['schema']}.
+        {cfg['sources']['zensus_population']['table']}""",
+        geom_col="geom_point",
+    )
+
+    gdf_boundaries = db.select_geodataframe(
+        f"""SELECT * FROM  {cfg['sources']['vg250_municipalities']['schema']}.
+        {cfg['sources']['vg250_municipalities']['table']}""",
+        geom_col="geometry",
+        epsg=3035,
+    )
+
+    # Join vg250 with zensus cells
+    join = gpd.sjoin(gdf, gdf_boundaries, how="inner", op="intersects")
+
+    # Deal with cells that don't interect with boundaries (e.g. at borders)
+    missing_cells = gdf[(~gdf.id.isin(join.id)) & (gdf.population > 0)]
+
+    # start with buffer
+    buffer = 0
+
+    # increase buffer until every zensus cell is matched to a nuts3 region
+    while len(missing_cells) > 0:
+        buffer += 100
+        boundaries_buffer = gdf_boundaries.copy()
+        boundaries_buffer.geometry = boundaries_buffer.geometry.buffer(buffer)
+        join_missing = gpd.sjoin(
+            missing_cells, boundaries_buffer, how="inner", op="intersects"
+        )
+        join = join.append(join_missing)
+        missing_cells = gdf[(~gdf.id.isin(join.id)) & (gdf.population > 0)]
+    print(f"Maximal buffer to match zensus points to vg250: {buffer}m")
+
+    # drop duplicates
+    join = join.drop_duplicates(subset=["id"])
+
+    # Insert results to database
+    join.rename(
+        {
+            "id": "zensus_population_id",
+            "geom_point": "zensus_geom",
+            "nuts": "vg250_nuts3",
+            "gid": "vg250_municipality_id",
+        },
+        axis=1,
+    )[
+        ["zensus_population_id", "zensus_geom",
+         "vg250_municipality_id", "vg250_nuts3"]
+    ].set_geometry(
+        "zensus_geom"
+    ).to_postgis(
+        cfg["targets"]["map"]["table"],
+        schema=cfg["targets"]["map"]["schema"],
+        con=local_engine,
+        if_exists="replace",
+    )
+
+
 def inside_germany():
     """
     Filter zensus data by data inside Germany and population > 0
@@ -145,6 +233,12 @@ def inside_germany():
         bind=engine_local_db, checkfirst=True
     )
 
+    # Selects zensus cells in German boundaries from vg250
+    cells_in_germany = db.select_dataframe(
+        """SELECT zensus_population_id
+        FROM boundaries.egon_map_zensus_vg250"""
+    ).zensus_population_id.values.tolist()
+
     with db.session_scope() as s:
         # Query relevant data from zensus population table
         q = (
@@ -156,12 +250,7 @@ def inside_germany():
                 DestatisZensusPopulationPerHa.geom,
             )
             .filter(DestatisZensusPopulationPerHa.population > 0)
-            .filter(
-                func.ST_Contains(
-                    func.ST_Transform(Vg250Sta.geometry, 3035),
-                    DestatisZensusPopulationPerHa.geom_point,
-                )
-            )
+            .filter(DestatisZensusPopulationPerHa.id.in_(cells_in_germany))
         )
 
         # Insert above queried data into new table
@@ -192,87 +281,36 @@ def population_in_municipalities():
 
     srid = 3035
 
-    # Prepare query from vg250 and zensus data
-    with db.session_scope() as session:
-        q = (
-            session.query(
-                Vg250Gem.gid.label("gid"),
-                Vg250Gem.gen.label("gen"),
-                Vg250Gem.bez,
-                Vg250Gem.bem,
-                Vg250Gem.nuts,
-                Vg250Gem.rs_0,
-                Vg250Gem.ags_0,
-                (
-                    func.ST_Area(func.ST_Transform(Vg250Gem.geometry, srid))
-                    / 10000
-                ).label(
-                    "area_ha"
-                ),  # ha
-                (
-                    func.ST_Area(func.ST_Transform(Vg250Gem.geometry, srid))
-                    / 1000000
-                ).label(
-                    "area_km2"
-                ),  # km
-                func.sum(
-                    func.coalesce(
-                        DestatisZensusPopulationPerHaInsideGermany.population,
-                        0,
-                    )
-                ).label("population_total"),
-                func.count(
-                    DestatisZensusPopulationPerHaInsideGermany.geom
-                ).label("cell_count"),
-                func.coalesce(
-                    func.sum(
-                        func.coalesce(
-                            DestatisZensusPopulationPerHaInsideGermany.population,
-                            0,
-                        )
-                    )
-                    / (
-                        func.ST_Area(
-                            func.ST_Transform(Vg250Gem.geometry, srid)
-                        )
-                        / 1000000
-                    ),
-                    0,
-                ).label("population_density"),
-                func.ST_Transform(Vg250Gem.geometry, srid).label("geom"),
-            )
-            .filter(
-                func.ST_Contains(
-                    func.ST_Transform(Vg250Gem.geometry, srid),
-                    DestatisZensusPopulationPerHaInsideGermany.geom_point,
-                )
-            )
-            .group_by(Vg250Gem.gid)
-        )
+    gem = db.select_geodataframe(
+        "SELECT * FROM boundaries.vg250_gem",
+        geom_col="geometry",
+        epsg=srid,
+        index_col='gid'
+    )
 
-        # Insert spatially joined data
-        insert = Vg250GemPopulation.__table__.insert().from_select(
-            [
-                Vg250GemPopulation.gid,
-                Vg250GemPopulation.gen,
-                Vg250GemPopulation.bez,
-                Vg250GemPopulation.bem,
-                Vg250GemPopulation.nuts,
-                Vg250GemPopulation.ags_0,
-                Vg250GemPopulation.rs_0,
-                Vg250GemPopulation.area_ha,
-                Vg250GemPopulation.area_km2,
-                Vg250GemPopulation.population_total,
-                Vg250GemPopulation.cell_count,
-                Vg250GemPopulation.population_density,
-                Vg250GemPopulation.geom,
-            ],
-            q,
-        )
+    gem["area_ha"] = gem.area / 10000
 
-        # Execute and commit (trigger transactions in database)
-        session.execute(insert)
-        session.commit()
+    gem["area_km2"] = gem.area / 1000000
+
+    population = db.select_dataframe(
+        """SELECT id, population, vg250_municipality_id
+        FROM society.destatis_zensus_population_per_ha
+        INNER JOIN boundaries.egon_map_zensus_vg250 ON (
+             society.destatis_zensus_population_per_ha.id =
+             boundaries.egon_map_zensus_vg250.zensus_population_id)
+        WHERE population > 0"""
+    )
+
+    gem["population_total"] = population.groupby(
+        'vg250_municipality_id').population.sum().fillna(0)
+
+    gem["cell_count"] = population.groupby(
+        'vg250_municipality_id').population.count()
+
+    gem["population_density"] = gem["population_total"]/gem["area_km2"]
+
+    gem.reset_index().to_postgis("vg250_gem_population", schema= "boundaries",
+                   con=db.engine(), if_exists='replace')
 
 
 def add_metadata_zensus_inside_ger():
@@ -299,37 +337,37 @@ def add_metadata_zensus_inside_ger():
         "sources": [
             {
                 "name": "Statistisches Bundesamt (Destatis) - Ergebnisse des "
-                        "Zensus 2011 zum Download",
+                "Zensus 2011 zum Download",
                 "description": "Als Download bieten wir Ihnen auf dieser Seite "
-                               "zusätzlich zur Zensusdatenbank CSV- und "
-                               "teilweise Excel-Tabellen mit umfassenden "
-                               "Personen-, Haushalts- und Familien- sowie "
-                               "Gebäude- und Wohnungs­merkmalen. Die "
-                               "Ergebnisse liegen auf Bundes-, Länder-, Kreis- "
-                               "und Gemeinde­ebene vor. Außerdem sind einzelne "
-                               "Ergebnisse für Gitterzellen verfügbar.",
+                "zusätzlich zur Zensusdatenbank CSV- und "
+                "teilweise Excel-Tabellen mit umfassenden "
+                "Personen-, Haushalts- und Familien- sowie "
+                "Gebäude- und Wohnungs­merkmalen. Die "
+                "Ergebnisse liegen auf Bundes-, Länder-, Kreis- "
+                "und Gemeinde­ebene vor. Außerdem sind einzelne "
+                "Ergebnisse für Gitterzellen verfügbar.",
                 "url": "https://www.zensus2011.de/SharedDocs/Aktuelles/Ergebnis"
-                       "se/DemografischeGrunddaten.html;jsessionid=E0A2B4F894B2"
-                       "58A3B22D20448F2E4A91.2_cid380?nn=3065474",
+                "se/DemografischeGrunddaten.html;jsessionid=E0A2B4F894B2"
+                "58A3B22D20448F2E4A91.2_cid380?nn=3065474",
                 "license": "",
                 "copyright": "© Statistische Ämter des Bundes und der Länder 2014",
             },
             {
                 "name": "Dokumentation - Zensus 2011 - Methoden und Verfahren",
                 "description": "Diese Publikation beschreibt ausführlich die "
-                               "Methoden und Verfahren des registergestützten "
-                               "Zensus 2011; von der Datengewinnung und "
-                               "-aufbereitung bis hin zur Ergebniserstellung"
-                               " und Geheimhaltung. Der vorliegende Band wurde "
-                               "von den Statistischen Ämtern des Bundes und "
-                               "der Länder im Juni 2015 veröffentlicht.",
+                "Methoden und Verfahren des registergestützten "
+                "Zensus 2011; von der Datengewinnung und "
+                "-aufbereitung bis hin zur Ergebniserstellung"
+                " und Geheimhaltung. Der vorliegende Band wurde "
+                "von den Statistischen Ämtern des Bundes und "
+                "der Länder im Juni 2015 veröffentlicht.",
                 "url": "https://www.destatis.de/DE/Publikationen/Thematisch/Be"
-                       "voelkerung/Zensus/ZensusBuLaMethodenVerfahren51211051"
-                       "19004.pdf?__blob=publicationFile",
+                "voelkerung/Zensus/ZensusBuLaMethodenVerfahren51211051"
+                "19004.pdf?__blob=publicationFile",
                 "license": "Vervielfältigung und Verbreitung, auch "
-                           "auszugsweise, mit Quellenangabe gestattet.",
+                "auszugsweise, mit Quellenangabe gestattet.",
                 "copyright": "© Statistisches Bundesamt, Wiesbaden, 2015 "
-                             "(im Auftrag der Herausgebergemeinschaft)",
+                "(im Auftrag der Herausgebergemeinschaft)",
             },
         ],
         "license": {
@@ -338,18 +376,18 @@ def add_metadata_zensus_inside_ger():
             "version": "2.0",
             "url": "www.govdata.de/dl-de/by-2-0",
             "instruction": "Empfohlene Zitierweise des Quellennachweises: "
-                           "Datenquelle: Statistisches Bundesamt, Wiesbaden, "
-                           "Genesis-Online, <optional> Abrufdatum; Datenlizenz "
-                           "by-2-0. Quellenvermerk bei eigener Berechnung / "
-                           "Darstellung: Datenquelle: Statistisches Bundesamt, "
-                           "Wiesbaden, Genesis-Online, <optional> Abrufdatum; "
-                           "Datenlizenz by-2-0; eigene Berechnung/eigene "
-                           "Darstellung. In elektronischen Werken ist im "
-                           "Quellenverweis dem Begriff (Datenlizenz by-2-0) "
-                           "der Link www.govdata.de/dl-de/by-2-0 als "
-                           "Verknüpfung zu hinterlegen.",
+            "Datenquelle: Statistisches Bundesamt, Wiesbaden, "
+            "Genesis-Online, <optional> Abrufdatum; Datenlizenz "
+            "by-2-0. Quellenvermerk bei eigener Berechnung / "
+            "Darstellung: Datenquelle: Statistisches Bundesamt, "
+            "Wiesbaden, Genesis-Online, <optional> Abrufdatum; "
+            "Datenlizenz by-2-0; eigene Berechnung/eigene "
+            "Darstellung. In elektronischen Werken ist im "
+            "Quellenverweis dem Begriff (Datenlizenz by-2-0) "
+            "der Link www.govdata.de/dl-de/by-2-0 als "
+            "Verknüpfung zu hinterlegen.",
             "copyright": "Statistisches Bundesamt, Wiesbaden, Genesis-Online; "
-                         "Datenlizenz by-2-0; eigene Berechnung",
+            "Datenlizenz by-2-0; eigene Berechnung",
         },
         "contributors": [
             {
@@ -432,7 +470,7 @@ def add_metadata_vg250_gem_pop():
 
     metadata = {
         "title": "Municipalities (BKG Verwaltungsgebiete 250) and population "
-                 "(Destatis Zensus)",
+        "(Destatis Zensus)",
         "description": "Municipality data enriched by population data",
         "language": ["DE"],
         "spatial": {